requirements.txt
# Core Data Science Libraries
pandas==2.2.0
numpy==1.26.3
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0

# Data Quality & Validation
great-expectations==0.18.8
pandera==0.18.0

# Database & Storage
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pyarrow==14.0.2

# Configuration & Environment
pydantic==2.5.3
python-dotenv==1.0.0
pyyaml==6.0.1

# Logging & Monitoring
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-cov==4.1.0

# Jupyter & Interactive Development
jupyter==1.0.0
ipykernel==6.28.0
ipywidgets==8.1.1

# Code Quality
black==23.12.1
flake8==7.0.0
mypy==1.8.0
